tratar erros

+++ src/server.ts
@@ -20,39 +20,101 @@ routes.get("/", (req: Request, res: Response) => {
 });
 
 routes.get("/getAllProducts", async (req: Request, res: Response) => {
-  await redis.synchronize();
-  const result = await redis.searchAllProducts();
+  try {
+    await redis.synchronize();
 
-  const products = result.documents;
-  res.statusCode = 200;
-  res.type("application/json");
-  res.send(products);
+    const result = await redis.searchAllProducts();
+
+    if (result.total === 0) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Products not found." });
+      return;
+    }
+
+    const products = result.documents;
+    res.statusCode = 200;
+    res.type("application/json");
+    res.send(products);
+  } catch (error) {
+    res.statusCode = 500;
+    res.type("application/json");
+    res.send({
+      error: error instanceof Error ? error.message : "Internal Server Error",
+    });
+    return;
+  }
 });
 
 routes.get("/getById/:id", async (rec: Request, res: Response) => {
-  await redis.synchronize();
+  try {
+    await redis.synchronize();
 
-  const id = rec.params.id as unknown as number;
-  const result = await redis.searchProductById(id);
+    const idParam = rec.params.id as unknown as number;
 
-  const product = JSON.stringify(result.documents[0], null, 2);
+    const id = Number(idParam);
+    if (!Number.isInteger(id)) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Invalid id." });
+      return;
+    }
 
-  res.statusCode = 200;
-  res.type("application/json");
-  res.send(product);
+    const result = await redis.searchProductById(id);
+    if (result.total === 0) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Product not found." });
+      return;
+    }
+
+    const product = JSON.stringify(result.documents[0], null, 2);
+
+    res.statusCode = 200;
+    res.type("application/json");
+    res.send(product);
+  } catch (error) {
+    res.statusCode = 500;
+    res.type("application/json");
+    res.send({
+      error: error instanceof Error ? error.message : "Internal Server Error",
+    });
+    return;
+  }
 });
 
 routes.post("/create", async (rec: Request, res: Response) => {
-  await redis.synchronize();
+  try {
+    await redis.synchronize();
 
-  const newProduct = rec.body as Product;
-  const products = await productsRepo.create(newProduct);
+    const newProduct = rec.body as Product;
+    if (
+      !newProduct ||
+      typeof newProduct.name !== "string" ||
+      typeof newProduct.price !== "number" ||
+      typeof newProduct.description !== "string"
+    ) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Invalid product data." });
+      return;
+    }
 
-  redis.create(products);
+    const products = await productsRepo.create(newProduct);
 
-  res.statusCode = 200;
-  res.type("application/json");
-  res.send(products);
+    redis.create(products);
+
+    res.statusCode = 200;
+    res.type("application/json");
+    res.send(products);
+  } catch (e) {
+    res.statusCode = 500;
+    res.type("application/json");
+    res.send({
+      error: e instanceof Error ? e.message : "Internal Server Error",
+    });
+    return;
+  }
 });
 
 routes.post("/update", async (rec: Request, res: Response) => {
@@ -61,42 +123,78 @@ routes.post("/update", async (rec: Request, res: Response) => {
 
     const updatedProduct = rec.body as Product;
 
-    if (!updatedProduct) {
+    if (
+      !updatedProduct ||
+      typeof updatedProduct.id !== "number" ||
+      typeof updatedProduct.name !== "string" ||
+      typeof updatedProduct.price !== "number" ||
+      typeof updatedProduct.description !== "string"
+    ) {
       res.statusCode = 400;
       res.type("application/json");
       res.send({ error: "Invalid product data." });
+      return;
     }
 
     const products = await productsRepo.update(updatedProduct);
 
     if (!products) {
       res.statusCode = 404;
-      res.type("application/json")
+      res.type("application/json");
       res.send({ error: "Product not found." });
+      return;
     }
 
-    redis.update(products);
+    await redis.update(products);
 
     res.statusCode = 200;
     res.type("application/json");
     res.send(products);
   } catch (e) {
-    res.statusCode = 404;
-    res.send({ error: e})
+    res.statusCode = 500;
+    res.type("application/json");
+    res.send({
+      error: e instanceof Error ? e.message : "Internal Server Error",
+    });
+    return;
   }
 });
 
 routes.delete("/delete/:id", async (rec: Request, res: Response) => {
-  await redis.synchronize();
-  const id = rec.params.id as unknown as number;
+  try {
+    await redis.synchronize();
 
-  await productsRepo.delete(id);
+    const idParam = rec.params.id as unknown as number;
 
-  redis.delete(id);
+    const id = Number(idParam);
+    if (!Number.isInteger(id)) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Invalid id." });
+      return;
+    }
 
-  res.statusCode = 200;
-  res.type("application/json");
-  res.send("OK");
+    const idResponse = await productsRepo.delete(id);
+    if (idResponse === 0) {
+      res.statusCode = 404;
+      res.type("application/json");
+      res.send({ error: "Product not found." });
+      return;
+    }
+
+    redis.delete(id);
+
+    res.statusCode = 200;
+    res.type("application/json");
+    res.send("OK");
+  } catch (e) {
+    res.statusCode = 500;
+    res.type("application/json");
+    res.send({
+      error: e instanceof Error ? e.message : "Internal Server Error",
+    });
+    return;
+  }
 });
 
 // aplicar as rotas na aplicação web backend.
